02-Working-with-libraries/02-libraries-for-development/08-networking/requirements.txt
dnspython
scapy